Move materials and logos overrides into funding router
