+++ controllers/funding/materials/lib/normalise-user-input.js
