+++ controllers/funding/materials/views/order-email.njk
@@ -1,4 +1,4 @@
-{% extends "views/layouts/emails.njk" %}
+{% extends "layouts/emails.njk" %}
 
 {% block content %}
     {{ copy | safe }}
